backend/upload-video/requirements.txt
boto3==1.34.21
psycopg2-binary==2.9.9